requirements.txt
dash
dash_bootstrap_components
dash_daq
dash_auth
numpy==1.24.4
pandas==2.0.3
pyyaml==6.0.1
rospkg==1.5.0
wrapt
watchdog
openstacksdk